requirements.txt
connexion == 2.6.0
python_dateutil == 2.6.0
setuptools >= 21.0.0

pyjwt >= 2.1.0
flask-cors >= 3.0.10
Unicode >= 2.8
Unidecode >= 1.3.2
pymongo[srv] >= 3.12.0
Image >= 1.5.33
cryptography >= 3.4.8
requests >= 2.26.0